move DataBindingViewHolder to androidx-recyclerview module
