+++ gto-support-androidx-recyclerview/build.gradle.kts
@@ -7,4 +7,8 @@ configureAndroidLibrary()
 
 dependencies {
     api(libs.androidx.recyclerview)
+
+    // region *DataBindingAdapter
+    compileOnly(libs.androidx.databinding.runtime)
+    // endregion *DataBindingAdapter
 }
