+++ gto-support-androidx-recyclerview/src/main/kotlin/org/ccci/gto/android/common/androidx/recyclerview/adapter/DataBindingViewHolder.kt
@@ -1,0 +1,6 @@
+package org.ccci.gto.android.common.androidx.recyclerview.adapter
+
+import androidx.databinding.ViewDataBinding
+import androidx.recyclerview.widget.RecyclerView
+
+open class DataBindingViewHolder<B : ViewDataBinding>(val binding: B) : RecyclerView.ViewHolder(binding.root)
