+++ gto-support-recyclerview/src/main/kotlin/org/ccci/gto/android/common/recyclerview/adapter/DataBindingViewHolder.kt
@@ -1,6 +1,7 @@
 package org.ccci.gto.android.common.recyclerview.adapter
 
 import androidx.databinding.ViewDataBinding
-import androidx.recyclerview.widget.RecyclerView
+import org.ccci.gto.android.common.androidx.recyclerview.adapter.DataBindingViewHolder
 
-open class DataBindingViewHolder<B : ViewDataBinding>(val binding: B) : RecyclerView.ViewHolder(binding.root)
+@Deprecated("Since v3.11.2, use DataBindingViewHolder from gto-support-androidx-recyclerview instead.")
+open class DataBindingViewHolder<B : ViewDataBinding>(binding: B) : DataBindingViewHolder<B>(binding)
